'派单取消'

+++ clms-traffic/clms-traffic-service/src/main/java/com/lcdt/traffic/service/impl/SplitGoodsServiceImpl.java
@@ -452,7 +452,8 @@ public class SplitGoodsServiceImpl implements SplitGoodsService {
 
         boolean opFlag = false;
         for (SplitGoodsDetail obj : splitGoodsDetailList) {
-            if (obj.getSplitGoodsDetailId().equals(splitGoodsDetailId)) {
+
+            if (obj.getSplitGoodsDetailId().equals(splitGoodsDetailId)) { //要取消的派单数
                 for (PlanDetail obj1: waybillPlan.getPlanDetailList()) {
                         if (obj.getPlanDetailId().equals(obj1.getPlanDetailId())) {
                             obj1.setRemainderAmount(obj1.getRemainderAmount() + obj.getRemainAmount());//计划剩余数量=计划现剩余数量+派单剩余数量
@@ -461,7 +462,9 @@ public class SplitGoodsServiceImpl implements SplitGoodsService {
                             obj1.setUpdateName(user.getRealName());
                             planDetailMapper.updateByPrimaryKey(obj1);
                             opFlag = true;
-                            splitGoodsDetailMapper.deleteByPrimaryKey(splitGoodsDetailId);  //先删除子明细
+                            if((obj.getAllotAmount()-obj.getRemainAmount())==0) {
+                                splitGoodsDetailMapper.deleteByPrimaryKey(splitGoodsDetailId);  //先删除子明细
+                            }
                             break;
                         }
                 }
